+++ base/src/sign.rs
@@ -42,6 +42,7 @@ pub trait UnsignedAbs {
 /// assert!(5.abs_eq(&-5));
 /// assert!(12.3.abs_eq(&-12.3));
 /// ```
+#[deprecated(since = "0.5", note = "AbsEq will be moved in AbsOrd in v0.5")] // TODO(v0.5): deprecate
 pub trait AbsEq<Rhs = Self> {
     fn abs_eq(&self, rhs: &Rhs) -> bool;
 }