+++ rational/tests/cmp.rs
@@ -1,10 +1,15 @@
 #![allow(clippy::nonminimal_bool)]
 
+use dashu_base::AbsOrd;
+use dashu_float::{DBig, FBig};
 use dashu_ratio::{RBig, Relaxed};
 use std::{
+    cmp::Ordering,
     collections::hash_map::DefaultHasher,
     hash::{Hash, Hasher},
 };
+
+type FBin = FBig;
 
 mod helper_macros;
 
@@ -104,4 +109,96 @@ fn test_cmp() {
     assert!(rbig!(-113 / 36) > rbig!(-355 / 113));
     assert!(rbig!(~-22/7) < rbig!(~-355/113));
     assert!(rbig!(~-113/36) > rbig!(~-355/113));
+}
+
+#[test]
+fn test_abs_ord() {
+    let ubig_cases = [
+        (rbig!(0), ubig!(0), Ordering::Equal),
+        (rbig!(0), ubig!(1), Ordering::Less),
+        (rbig!(1), ubig!(1), Ordering::Equal),
+        (rbig!(-1), ubig!(1), Ordering::Equal),
+        (rbig!(1 / 10), ubig!(0), Ordering::Greater),
+        (rbig!(1 / 10), ubig!(1), Ordering::Less),
+        (rbig!(-1 / 10), ubig!(0), Ordering::Greater),
+        (rbig!(-1 / 10), ubig!(1), Ordering::Less),
+        (rbig!(999999 / 100), ubig!(9999), Ordering::Greater),
+        (rbig!(999999 / 100), ubig!(10000), Ordering::Less),
+        (rbig!(-999999 / 100), ubig!(9999), Ordering::Greater),
+        (rbig!(-999999 / 100), ubig!(10000), Ordering::Less),
+    ];
+
+    for (r, i, order) in ubig_cases {
+        assert_eq!(r.abs_cmp(&i), order);
+        assert_eq!(i.abs_cmp(&r), order.reverse());
+    }
+
+    let ibig_cases = [
+        (rbig!(0), ibig!(0), Ordering::Equal),
+        (rbig!(0), ibig!(1), Ordering::Less),
+        (rbig!(0), ibig!(-1), Ordering::Less),
+        (rbig!(1), ibig!(1), Ordering::Equal),
+        (rbig!(1), ibig!(-1), Ordering::Equal),
+        (rbig!(-1), ibig!(1), Ordering::Equal),
+        (rbig!(-1), ibig!(-1), Ordering::Equal),
+        (rbig!(1 / 10), ibig!(0), Ordering::Greater),
+        (rbig!(1 / 10), ibig!(1), Ordering::Less),
+        (rbig!(-1 / 10), ibig!(0), Ordering::Greater),
+        (rbig!(-1 / 10), ibig!(-1), Ordering::Less),
+        (rbig!(-999999 / 100), ibig!(9999), Ordering::Greater),
+        (rbig!(-999999 / 100), ibig!(10000), Ordering::Less),
+        (rbig!(-999999 / 100), ibig!(-9999), Ordering::Greater),
+        (rbig!(-999999 / 100), ibig!(-10000), Ordering::Less),
+    ];
+
+    for (r, i, order) in ibig_cases {
+        assert_eq!(r.abs_cmp(&i), order);
+        assert_eq!(i.abs_cmp(&r), order.reverse());
+    }
+
+    let fbig_cases = [
+        (rbig!(0), FBin::ZERO, Ordering::Equal),
+        (rbig!(0), FBin::ONE, Ordering::Less),
+        (rbig!(0), FBin::NEG_ONE, Ordering::Less),
+        (rbig!(1), FBin::ONE, Ordering::Equal),
+        (rbig!(1), FBin::NEG_ONE, Ordering::Equal),
+        (rbig!(-1), FBin::ONE, Ordering::Equal),
+        (rbig!(-1), FBin::NEG_ONE, Ordering::Equal),
+        (rbig!(1 / 2), FBin::from_str_native("-0x1p-1").unwrap(), Ordering::Equal),
+        (rbig!(-9 / 2), FBin::from_str_native("-0x9p-1").unwrap(), Ordering::Equal),
+        (rbig!(-1 / 1024), FBin::from_str_native("0x1p-10").unwrap(), Ordering::Equal),
+        (
+            rbig!(1 / 1267650600228229401496703205376),
+            FBin::from_str_native("0x1p-100").unwrap(),
+            Ordering::Equal,
+        ),
+    ];
+
+    for (r, f, order) in fbig_cases {
+        assert_eq!(r.abs_cmp(&f), order);
+        assert_eq!(f.abs_cmp(&r), order.reverse());
+    }
+
+    let dbig_cases = [
+        (rbig!(0), DBig::ZERO, Ordering::Equal),
+        (rbig!(0), DBig::ONE, Ordering::Less),
+        (rbig!(0), DBig::NEG_ONE, Ordering::Less),
+        (rbig!(1), DBig::ONE, Ordering::Equal),
+        (rbig!(1), DBig::NEG_ONE, Ordering::Equal),
+        (rbig!(-1), DBig::ONE, Ordering::Equal),
+        (rbig!(-1), DBig::NEG_ONE, Ordering::Equal),
+        (rbig!(1 / 10), DBig::from_str_native("-0.1").unwrap(), Ordering::Equal),
+        (rbig!(-11 / 10), DBig::from_str_native("-1.1").unwrap(), Ordering::Equal),
+        (rbig!(-1 / 9765625), DBig::from_str_native("1.024e-7").unwrap(), Ordering::Equal),
+        (
+            rbig!(1 / 7888609052210118054117285652827862296732064351090230047702789306640625),
+            DBig::from_str_native("1.267650600228229401496703205376e-70").unwrap(),
+            Ordering::Equal,
+        ),
+    ];
+
+    for (r, d, order) in dbig_cases {
+        assert_eq!(r.abs_cmp(&d), order, "{}, {}", r, d);
+        assert_eq!(d.abs_cmp(&r), order.reverse());
+    }
 }